Create annotated git tags.
Lerna uses its git tags to determine which packages have been
updated since the last publication, so it's important that tags be
committed to the repository. Git's annotated tags are designed
for this in a way that lightweight tags are not. For example, only
annotated tags are pushed by `git push --follow-tags`.